+++ modules/system/database/migrations/2013_10_01_000001_Db_Deferred_Bindings.php
@@ -17,7 +17,7 @@ class DbDeferredBindings extends Migration
             $table->string('slave_type')->index();
             $table->string('slave_id')->index();
             $table->string('session_key');
-            $table->boolean('bind')->default(true);
+            $table->boolean('is_bind')->default(true);
             $table->timestamps();
         });
     }