+++ modules/system/database/migrations/2013_10_01_000008_Db_System_Email_Templates.php
@@ -1,0 +1,31 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class DbSystemEmailTemplates extends Migration
+{
+
+    public function up()
+    {
+        Schema::create('system_email_templates', function(Blueprint $table)
+        {
+            $table->engine = 'InnoDB';
+            $table->increments('id');
+            $table->string('code')->nullable();
+            $table->string('subject')->nullable();
+            $table->text('description')->nullable();
+            $table->text('content_html')->nullable();
+            $table->text('content_text')->nullable();
+            $table->integer('layout_id')->index()->nullable();
+            $table->boolean('is_custom')->default(0);
+            $table->timestamps();
+        });
+    }
+
+    public function down()
+    {
+        Schema::drop('system_email_templates');
+    }
+
+}